fix(tests): fold the stdout tests into the mirror file (PS-204)
The audit requires a test file to mirror its src module 1:1, in the
mirrored directory. test__helpers_stdout_only.py was wrong twice: it
implied a src module of that name, and it sat in tests/_mcp/ while
_helpers.py lives in _mcp/_tools/.

Folded into the existing tests/_mcp/_tools/test__helpers.py, which is
the correct mirror and already exercises the same two functions.

133 passed, 1 skipped in that directory.
